farm_quest_django/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
mysqlclient
pandas
matplotlib
plotly
seaborn
tensorflow
opencv-python
ultralytics
environ